+++ js/sdk/index.mjs
@@ -1,6 +1,6 @@
 /**
  * @file Exports the Portal SDK as an ES6 module
  */
-import { SDK, Store } from './index.js'
+import { SDK, IndexedDB } from './index.js'
 export default SDK
-export { SDK, Store }
+export { SDK, IndexedDB }
